:memo: technical-assistance: Licensing, repositories, community
This commit restructures the technical assistance page to change from a
"repositories, forums, licensing" structure to "licensing, repositories,
community." Licensing is moved to the front because it is often an early
business requirement that ties to the business model of a project.
Repositories gains a new note about creating documentation. Forums are
reframed as community and community-building, since collaboration
happens in many places.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>