Update biz-models.md
Before considering if it is a Digital Public Good (DPG), we have to understand, what is considered open source?
 - This question isn't clear. Are we saying before we can review if something is a DPG we are asking are they open source?

Are we saying startups must be open source or their work should be? I suggest rereading this page and updating where needed to be specific. 

Should it be myth and then fact? Truth sounds like the myth is true (which it isn't!).